use host.GetGraph to get pipelines info for zpages
