--- service/internal/zpages/templates/component_header.html
@@ -1,6 +1,0 @@
-{{$link := .Link}}
-{{- if $link -}}
-    <h6><a href="{{.ComponentEndpoint}}">{{.Name}}</a></h6>
-{{- else -}}
-    <h6>{{.Name}}</h6>
-{{- end -}}
