--- service/internal/zpages/templates/page_footer.html
@@ -1,2 +1,0 @@
-</body>
-</html>
