--- service/internal/zpages/templates/page_header.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en"><head>
-    <meta charset="utf-8">
-    <title>{{.Title}}</title>
-    <link rel="shortcut icon" href="https://opentelemetry.io/favicons/favicon.ico"/>
-    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-    <link rel="stylesheet" href="https://code.getmdl.io/1.3.0/material.indigo-pink.min.css">
-    <script defer src="https://code.getmdl.io/1.3.0/material.min.js"></script>
-</head>
-<body>
-<h2>{{.Title}}</h2>
